Change the FromRow instances to applicative style
The concerns detailed in 5f045cc5c80d05b693a3d85a0f7634f7c3bcb471 should
now be taken care of by the revised interface.

Namely, the Ok constructor must be forced in order to determine whether
or not an exception is thrown,  and the Ok constructor evaluates it's
argument to whnf.