Clarify loImport/transaction issue in docs
As it turns out,  c_lo_creat, etc. are just wrappers around PQfn to
the corresponsing server-side functions.

Also, there is no non-blocking variant of PQfn in libpq,  though a
protocol-level client could certainly support fast-path calls with a
non-blocking implementation.

Thus as postgresql-simple moves to a non-blocking implementation underneath
the library,  it will make sense to eventually abandon the large object
binding in postgresql-libpq in favor of non-blocking calls to execParams.