initial support for using foundationdb
FoundationDB is a modern database
with a focus on perfecting distributed transactions
and fault tolerance (the hard stuff).
On top of that, they build layers, including
a [SQL Layer](https://foundationdb.com/layers/sql).
The SQL Layer speaks the PostgresSQL protocol and
works with libpq.

This patch adds a foundatbiondb flag that sets up CPP.
It CPP removes some connection initialization code
and code to set the isolation level.
Isolation levels are not needed in FoundationDB
because every transaction has a
[serializeable isolation level](http://blog.foundationdb.com/whiteboard-series-isolation).

most of the test cases are PostgresSQL specific
and thus are removed with CPP.
The few that are left show that
binary serialization definitely has issues.
There are probably some other similar issues to fix around data types.