Fix TypeInfo for the _record type
This issue was exposed in #153.

I don't like the fact that this change means the information in the
static table doesn't entirely agree with what would otherwise be fetched
dynamically by the Database.PostgreSQL.Simple.TypeInfo module, but
this does appear to be a special case,  and handling it here does seem
to be less worse then handling it there.

Plus,  now GenTypeInfo should be able to deal with array types,  so
we probably should strongly consider moving at least some of them into
the static table.