Minor cleanups relating to #149
I did investigate using QuickCheck to generate round-trip test cases,
but I ran into minor issues regarding integrating the two testing libraries
into a single test suite.  And honestly, as-is,  the test cases as-is
are probably a reasonably complete test of the new Generic instances for
FromField and ToField.   In all likelihood,  the logic that
QuickCheck-generated roundtrips would really exercise is the FromField
and ToField instances.  (Which admittedly, is a major hole in the current
test suite.)