Rewrite the documentation in the Time module
That documentation has been barely touched, if at all,  in the three
years since I wrote it.   It was kinda painful to read,  as I've
learned a lot about timekeeping in the meantime.

This should be massively more informative and all-around better.
Perhaps it could be a bit more concise,  but this is a lengthy topic.