PostgreSQL wants type information for `ARRAY[]`
While PostgreSQL does not require type information for the `ARRAY[]`
syntax when it is non-empty, empty arrays result in an error:

```
SqlError
  { sqlState = "42P18", sqlExecStatus = FatalError
  , sqlErrorMsg = "cannot determine type of empty array", sqlErrorDetail = ""
  , sqlErrorHint =
    "Explicitly cast to the desired type, for example ARRAY[]::integer[]."
  }
```

i.e. `ARRAY[1,2,3,4]` is correctly interpreted to be `integer[]`, but
`ARRAY[]` does not.

The syntax `'{}'` does not require type information when empty (for some
reason). Unfortunately, that syntax would require manual escaping for
non-empty arrays (e.g. if the elements have type `text`).

This commit simply special-cases the `ToField` instance of `PGArray`
when the array is empty to use the `{}` syntax.