Copy: Show error message on failure
This ensures that the exception thrown includes the cause. Sadly it
seems that the only way to determine that a COPY IN has failed is to
fail to parse the row count. It seems that all of the getResults calls
after the putCopyEnd command inexplicably return success, despite the
fact that something failed.