HStore implementation to parse NULL values and filter them out from the HStoreList.
This has a minor side-effect where hstores with NULL values will not really
round-trip properly, i.e. if we read the following from the DB

    "foo" => "bar", "fooNull" => NULL

we will get the following in Haskell

    [("foo", "bar")]

and when stored back to the DB, it will get converted to:

   "foo" => "bar"

This results in minor loss in fidelity. However, for all practical purposes,
the absence of a key, and the key being assigned a NULL value is the same
thing. How many programs out there would really be doing different things based
on this condition?